requirements.txt
tweepy>=4.14.0
openai>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
